begin merging codemirror setup
